Move Assignments out of LazyEvaluator
